backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pymysql
python-dotenv
pydantic
pandas
openpyxl
python-multipart
